feat: 대화 타입 수정 로직 변경, 테스트 코드 변경(#100)

+++ src/main/java/org/example/tokpik_be/type/service/TopicTypeService.java
@@ -56,14 +56,16 @@ public class TopicTypeService {
             throw new GeneralException(TypeException.DUPLICATE_TYPES);
         }
 
+        List<TopicType> topicTypes = topicTypeRepository.findAllById(request.topicTypeIds());
+
+        if (topicTypes.size() != request.topicTypeIds().size()) {
+            throw new GeneralException(TypeException.TYPE_NOT_FOUND);
+        }
+
         userTopicTypeRepository.deleteByUserId(user.getId());
 
-        for (long typeId : request.topicTypeIds()) {
-            if (!topicTypeRepository.existsById(typeId)) {
-                throw new GeneralException(TypeException.TYPE_NOT_FOUND);
-            }
-            UserTopicType userTopicType = new UserTopicType(user.getId(),
-                topicTypeRepository.findById(typeId).get());
+        for (TopicType topicType : topicTypes) {
+            UserTopicType userTopicType = new UserTopicType(user.getId(), topicType);
             userTopicTypeRepository.save(userTopicType);
         }
